refactor(subagent): extract execution backends into subagent_execution.py
Split 8 internal functions (~475 lines) from the 1,625-line subagent.py
monolith into a focused subagent_execution.py module:

- _load_agent_memory, _build_memory_system_message (agent memory)
- _create_subagent_thread, _run_subagent_subprocess (spawn backends)
- _monitor_subprocess, _cleanup_isolation (process lifecycle)
- _summarize_result (result formatting)
- _run_planner (multi-subtask orchestration)

subagent.py reduced from 1,625 → 1,150 lines (29% reduction).
No public API changes — all symbols remain importable from subagent.py.
